+++ fixtures/ios_app_with_helpers/Projects/App/Project.swift
@@ -2,5 +2,5 @@ import ProjectDescription
 import ProjectDescriptionHelpers
 
 let project = Project.app(name: "App", platform: .iOS, dependencies: [
-    .project(target: "AppKit", path: .relativeToManifest("../AppKit"))
+    .project(target: "AppKit", path: "//Projects/AppKit")
 ])