+++ fixtures/ios_app_with_helpers/Projects/AppKit/Project.swift
@@ -2,5 +2,5 @@ import ProjectDescription
 import ProjectDescriptionHelpers
 
 let project = Project.framework(name: "AppKit", platform: .iOS, dependencies: [
-    .project(target: "AppSupport", path: .relativeToManifest("../AppSupport"))
+    .project(target: "AppSupport", path: "//Projects/AppSupport")
 ])